+++ sdk/core/azure_core/src/http/poller.rs
@@ -7,12 +7,14 @@ use crate::{
     error::{ErrorKind, ErrorResponse},
     http::{
         headers::{HeaderName, Headers},
-        Format, JsonFormat, Response, StatusCode,
+        policies::create_public_api_span,
+        Context, Format, JsonFormat, Response, StatusCode,
     },
     sleep,
     time::{Duration, OffsetDateTime},
+    tracing::{Span, SpanStatus},
 };
-use futures::{stream::unfold, Stream, StreamExt};
+use futures::{channel::oneshot, stream::unfold, Stream, StreamExt};
 use serde::Deserialize;
 use std::{
     convert::Infallible,
@@ -20,7 +22,8 @@ use std::{
     future::{Future, IntoFuture},
     pin::Pin,
     str::FromStr,
-    task::{Context, Poll},
+    sync::Arc,
+    task::{Context as TaskContext, Poll},
 };
 
 /// Default retry time for long-running operations if no retry-after header is present
@@ -28,6 +31,8 @@ use std::{
 /// This value is the same as the default used in other Azure SDKs e.g.,
 /// <https://github.com/Azure/azure-sdk-for-python/blob/azure-core_1.35.0/sdk/core/azure-core/azure/core/polling/base_polling.py#L586>
 const DEFAULT_RETRY_TIME: Duration = Duration::seconds(30);
+
+/// Minimum retry time for long-running operations
 const MIN_RETRY_TIME: Duration = Duration::seconds(1);
 
 /// Represents the state of a [`Poller`].
@@ -145,18 +150,40 @@ impl<'de> Deserialize<'de> for PollerStatus {
 }
 
 /// Options to create the [`Poller`].
-#[derive(Debug, Clone, Copy)]
-pub struct PollerOptions {
+#[derive(Debug, Clone)]
+pub struct PollerOptions<'a> {
+    /// Allows customization of the method call.
+    pub context: Context<'a>,
     /// The time to wait between polling intervals in absence of a `retry-after` header.
     ///
     /// The default is 30 seconds. The minimum time enforced by [`Poller::from_callback`] is 1 second.
-    pub frequency: Option<Duration>,
+    pub frequency: Duration,
 }
 
-impl Default for PollerOptions {
+impl Default for PollerOptions<'_> {
     fn default() -> Self {
         Self {
-            frequency: Some(DEFAULT_RETRY_TIME),
+            frequency: DEFAULT_RETRY_TIME,
+            context: Context::new(),
+        }
+    }
+}
+
+impl<'a> PollerOptions<'a> {
+    /// Converts these poller options into an owned form so they can outlive the current scope.
+    #[must_use]
+    pub fn into_owned(self) -> PollerOptions<'static> {
+        PollerOptions {
+            context: self.context.into_owned(),
+            frequency: self.frequency,
+        }
+    }
+
+    /// Creates a new owned instance of `PollerOptions` so it can outlive the current scope.
+    pub fn to_owned(&self) -> PollerOptions<'static> {
+        PollerOptions {
+            context: self.context.to_owned(),
+            frequency: self.frequency,
         }
     }
 }
@@ -174,7 +201,7 @@ pub enum PollerResult<M: StatusMonitor, N, F: Format = JsonFormat> {
         /// The HTTP response with the status monitor.
         response: Response<M, F>,
         /// The optional client-specified [`Duration`] to wait before polling again.
-        retry_after: Option<Duration>,
+        retry_after: Duration,
         /// The next link / continuation token.
         next: N,
     },
@@ -389,7 +416,7 @@ where
     /// let url = "https://example.com/my_operation".parse().unwrap();
     /// let mut req = Request::new(url, Method::Post);
     ///
-    /// let poller = Poller::from_callback(move |operation_url: PollerState<Url>| {
+    /// let poller = Poller::from_callback(move |operation_url: PollerState<Url>,  poller_options| {
     ///     // The callback must be 'static, so you have to clone and move any values you want to use.
     ///     let pipeline = pipeline.clone();
     ///     let api_version = api_version.clone();
@@ -406,7 +433,7 @@ where
     ///             .append_pair("api-version", &api_version);
     ///
     ///         let resp = pipeline
-    ///             .send(&Context::new(), &mut req, None)
+    ///             .send(&poller_options.context, &mut req, None)
     ///             .await?;
     ///         let (status, headers, body) = resp.deconstruct();
     ///         let result: OperationResult = json::from_json(&body)?;
@@ -419,7 +446,7 @@ where
     ///                 let operation_url = format!("https://example.com/operations/{}", result.id).parse()?;
     ///                 Ok(PollerResult::InProgress {
     ///                     response: resp,
-    ///                     retry_after: None,
+    ///                     retry_after: poller_options.frequency,
     ///                     next: operation_url
     ///                 })
     ///             }
@@ -445,20 +472,21 @@ where
     /// ```
     pub fn from_callback<
         #[cfg(not(target_arch = "wasm32"))] N: AsRef<str> + Send + 'static,
-        #[cfg(not(target_arch = "wasm32"))] Fun: Fn(PollerState<N>) -> Fut + Send + 'static,
+        #[cfg(not(target_arch = "wasm32"))] Fun: Fn(PollerState<N>, PollerOptions<'static>) -> Fut + Send + 'static,
         #[cfg(not(target_arch = "wasm32"))] Fut: Future<Output = crate::Result<PollerResult<M, N, F>>> + Send + 'static,
         #[cfg(target_arch = "wasm32")] N: AsRef<str> + 'static,
-        #[cfg(target_arch = "wasm32")] Fun: Fn(PollerState<N>) -> Fut + 'static,
+        #[cfg(target_arch = "wasm32")] Fun: Fn(PollerState<N>, PollerOptions<'static>) -> Fut + 'static,
         #[cfg(target_arch = "wasm32")] Fut: Future<Output = crate::Result<PollerResult<M, N, F>>> + 'static,
     >(
         make_request: Fun,
-        options: Option<PollerOptions>,
+        options: Option<PollerOptions<'static>>,
     ) -> Self
     where
         M: Send + 'static,
         M::Output: Send + 'static,
         M::Format: Send + 'static,
     {
+        let options = options.unwrap_or_default();
         let (stream, target) = create_poller_stream(make_request, options);
         Self {
             stream: Box::pin(stream),
@@ -493,7 +521,7 @@ where
 {
     type Item = crate::Result<Response<M, F>>;
 
-    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
+    fn poll_next(self: Pin<&mut Self>, cx: &mut TaskContext<'_>) -> Poll<Option<Self::Item>> {
         let state = self.project().stream.poll_next(cx);
         if let Poll::Ready(Some(Ok(ref response))) = state {
             check_status_code(response)?;
@@ -579,18 +607,38 @@ enum State<N> {
     Done,
 }
 
+/// The type of the oneshot channel sender for the target future.
+type TargetTransmitterType<'a, M> = (Pin<BoxedFuture<M>>, Option<Context<'a>>);
+
+/// Represents the state used for each iteration through the poller stream.
+struct StreamState<'a, M, N, Fun>
+where
+    M: StatusMonitor,
+{
+    /// The current polling state (Init, InProgress, or Done)
+    state: State<N>,
+    /// The callback function to make requests
+    make_request: Fun,
+    /// Optional channel sender for the target future
+    target_tx: Option<oneshot::Sender<TargetTransmitterType<'a, M>>>,
+    /// The poller options
+    options: PollerOptions<'a>,
+    /// Whether a span was added to the context
+    added_span: bool,
+}
+
 fn create_poller_stream<
     M,
     F: Format,
     #[cfg(not(target_arch = "wasm32"))] N: AsRef<str> + Send + 'static,
-    #[cfg(not(target_arch = "wasm32"))] Fun: Fn(PollerState<N>) -> Fut + Send + 'static,
+    #[cfg(not(target_arch = "wasm32"))] Fun: Fn(PollerState<N>, PollerOptions<'static>) -> Fut + Send + 'static,
     #[cfg(not(target_arch = "wasm32"))] Fut: Future<Output = crate::Result<PollerResult<M, N, F>>> + Send + 'static,
     #[cfg(target_arch = "wasm32")] N: AsRef<str> + 'static,
-    #[cfg(target_arch = "wasm32")] Fun: Fn(PollerState<N>) -> Fut + 'static,
+    #[cfg(target_arch = "wasm32")] Fun: Fn(PollerState<N>, PollerOptions<'static>) -> Fut + 'static,
     #[cfg(target_arch = "wasm32")] Fut: Future<Output = crate::Result<PollerResult<M, N, F>>> + 'static,
 >(
     make_request: Fun,
-    options: Option<PollerOptions>,
+    options: PollerOptions<'static>,
 ) -> (
     impl Stream<Item = crate::Result<Response<M, F>>> + 'static,
     BoxedFuture<M>,
@@ -600,33 +648,48 @@ where
     M::Output: Send + 'static,
     M::Format: Send + 'static,
 {
-    use futures::channel::oneshot;
-
     let (target_tx, target_rx) = oneshot::channel();
-    let frequency = options
-        .unwrap_or_default()
-        .frequency
-        .unwrap_or(DEFAULT_RETRY_TIME);
+
     assert!(
-        frequency >= MIN_RETRY_TIME,
+        options.frequency >= MIN_RETRY_TIME,
         "minimum polling frequency is 1 second"
     );
-
     let stream = unfold(
         // We flow the `make_request` callback through the state value to avoid cloning.
-        (State::Init, make_request, Some(target_tx)),
-        move |(state, make_request, target_tx)| async move {
-            let result = match state {
+        StreamState::<M, N, Fun> {
+            state: State::Init,
+            make_request,
+            target_tx: Some(target_tx),
+            options,
+            added_span: false,
+        },
+        move |mut poller_stream_state| async move {
+            let result = match poller_stream_state.state {
                 State::Init => {
-                    tracing::debug!("initial operation request");
-                    make_request(PollerState::Initial).await
+                    // At the very start of polling, create a span for the entire request, and attach it to the context
+                    let span =
+                        create_public_api_span(&poller_stream_state.options.context, None, None);
+                    if let Some(ref s) = span {
+                        poller_stream_state.added_span = true;
+                        poller_stream_state.options.context =
+                            poller_stream_state.options.context.with_value(s.clone());
+                    }
+                    (poller_stream_state.make_request)(
+                        PollerState::Initial,
+                        poller_stream_state.options.clone(),
+                    )
+                    .await
                 }
                 State::InProgress(n) => {
                     tracing::debug!(
                         "subsequent operation request to {:?}",
                         AsRef::<str>::as_ref(&n)
                     );
-                    make_request(PollerState::More(n)).await
+                    (poller_stream_state.make_request)(
+                        PollerState::More(n),
+                        poller_stream_state.options.clone(),
+                    )
+                    .await
                 }
                 State::Done => {
                     tracing::debug!("done");
@@ -634,7 +697,23 @@ where
                 }
             };
             let (item, next_state) = match result {
-                Err(e) => return Some((Err(e), (State::Done, make_request, target_tx))),
+                Err(e) => {
+                    if poller_stream_state.added_span {
+                        if let Some(span) =
+                            poller_stream_state.options.context.value::<Arc<dyn Span>>()
+                        {
+                            // Mark the span as an error with an appropriate description.
+                            span.set_status(SpanStatus::Error {
+                                description: e.to_string(),
+                            });
+                            span.set_attribute("error.type", e.kind().to_string().into());
+                            span.end();
+                        }
+                    }
+
+                    poller_stream_state.state = State::Done;
+                    return Some((Err(e), poller_stream_state));
+                }
                 Ok(PollerResult::InProgress {
                     response,
                     retry_after,
@@ -642,35 +721,87 @@ where
                 }) => {
                     // Note that test-proxy automatically adds a transform that zeroes an existing `after-retry` header during playback, so don't check at runtime:
                     // <https://github.com/Azure/azure-sdk-tools/blob/a80b559d7682891f36a491b73f52fcb679d40923/tools/test-proxy/Azure.Sdk.Tools.TestProxy/RecordingHandler.cs#L1175>
-                    let duration = retry_after.unwrap_or(frequency);
-
-                    tracing::trace!("retry poller in {}s", duration.whole_seconds());
-                    sleep(duration).await;
+                    tracing::trace!("retry poller in {}s", retry_after.whole_seconds());
+                    sleep(retry_after).await;
 
                     (Ok(response), State::InProgress(n))
                 }
+                // Note that we will normally never reach this state. The normal progression of the `make_request` callback is to return `Succeeded` with a target future,
+                // and then the stream yields the final response and transitions to `Done` state.
+                // The only time that the `make_request` callback will normally enter the `Done` state directly is if the LRO fails or is canceled.
                 Ok(PollerResult::Done { response }) => (Ok(response), State::Done),
                 Ok(PollerResult::Succeeded {
                     response,
                     target: get_target,
                 }) => {
                     // Send the target callback through the channel
-                    if let Some(tx) = target_tx {
-                        let _ = tx.send(get_target());
+                    if let Some(tx) = poller_stream_state.target_tx.take() {
+                        let _ = tx.send((
+                            get_target(),
+                            if poller_stream_state.added_span {
+                                Some(poller_stream_state.options.context.clone())
+                            } else {
+                                None
+                            },
+                        ));
                     }
-                    // Also yield the final status response and set target_tx to None since we consumed it
-                    return Some((Ok(response), (State::Done, make_request, None)));
+                    // Also yield the final status response
+                    poller_stream_state.state = State::Done;
+                    return Some((Ok(response), poller_stream_state));
                 }
             };
 
-            // Flow 'make_request' and target_tx through to avoid cloning
-            Some((item, (next_state, make_request, target_tx)))
+            // Update state and return
+            poller_stream_state.state = next_state;
+            Some((item, poller_stream_state))
         },
     );
 
     let target = Box::new(async move {
         match target_rx.await {
-            Ok(fut) => fut.await,
+            Ok(target_state) => {
+                // Await the target future to get the final response from the poller.
+                let res = target_state.0.await;
+                // If we added a span to the target, take the result of the final target future to finalize the span.
+
+                if let Some(ctx) = target_state.1 {
+                    match &res {
+                        Ok(response) => {
+                            // When the result is done, finalize the span. Note that we only do that if we created the span in the first place,
+                            // otherwise it is the responsibility of the caller to end their span.
+                            if let Some(span) = ctx.value::<Arc<dyn Span>>() {
+                                // 5xx status codes SHOULD set status to Error.
+                                // The description should not be set because it can be inferred from "http.response.status_code".
+                                if response.status().is_server_error() {
+                                    span.set_status(SpanStatus::Error {
+                                        description: "".to_string(),
+                                    });
+                                }
+                                if response.status().is_client_error()
+                                    || response.status().is_server_error()
+                                {
+                                    span.set_attribute(
+                                        "error.type",
+                                        response.status().to_string().into(),
+                                    );
+                                }
+
+                                span.end();
+                            }
+                        }
+                        Err(err) => {
+                            if let Some(span) = ctx.value::<Arc<dyn Span>>() {
+                                span.set_status(SpanStatus::Error {
+                                    description: err.to_string(),
+                                });
+                                span.set_attribute("error.type", err.kind().to_string().into());
+                                span.end();
+                            }
+                        }
+                    }
+                }
+                res
+            }
             Err(err) => Err(crate::Error::with_error(
                 ErrorKind::Other,
                 err,
@@ -687,11 +818,11 @@ pub fn get_retry_after(
     headers: &Headers,
     retry_headers: &[HeaderName],
     options: &PollerOptions,
-) -> Option<Duration> {
+) -> Duration {
     #[cfg_attr(feature = "test", allow(unused_mut))]
     let duration =
         crate::http::policies::get_retry_after(headers, OffsetDateTime::now_utc, retry_headers)
-            .or(options.frequency);
+            .unwrap_or(options.frequency);
 
     #[cfg(feature = "test")]
     {
@@ -701,17 +832,14 @@ pub fn get_retry_after(
         // we need to override the frequency for services which do not send back supported headers in their response.
         if matches!(headers.get_optional::<RecordingMode>(), Ok(Some(mode)) if mode == RecordingMode::Playback)
         {
-            match duration {
-                Some(duration) if duration > Duration::ZERO => {
-                    tracing::debug!(
-                        "overriding {}s poller retry in playback",
-                        duration.whole_seconds()
-                    );
-                }
-                _ => {}
+            if duration > Duration::ZERO {
+                tracing::debug!(
+                    "overriding {}s poller retry in playback",
+                    duration.whole_seconds()
+                );
             }
 
-            return Some(Duration::ZERO);
+            return Duration::ZERO;
         }
     }
 
@@ -826,7 +954,7 @@ mod tests {
         };
 
         let mut poller = Poller::from_callback(
-            move |_| {
+            move |_, _| {
                 let client = mock_client.clone();
                 async move {
                     let req = Request::new("https://example.com".parse().unwrap(), Method::Get);
@@ -841,7 +969,7 @@ mod tests {
                     match test_status.status() {
                         PollerStatus::InProgress => Ok(PollerResult::InProgress {
                             response,
-                            retry_after: Some(Duration::ZERO),
+                            retry_after: Duration::ZERO,
                             next: "",
                         }),
                         _ => Ok(PollerResult::Done { response }),
@@ -906,9 +1034,8 @@ mod tests {
                 .boxed()
             }))
         };
-
         let mut poller = Poller::from_callback(
-            move |_| {
+            move |_, _| {
                 let client = mock_client.clone();
                 async move {
                     let req = Request::new("https://example.com".parse().unwrap(), Method::Get);
@@ -926,7 +1053,7 @@ mod tests {
                     match test_status.status() {
                         PollerStatus::InProgress => Ok(PollerResult::InProgress {
                             response,
-                            retry_after: Some(Duration::ZERO),
+                            retry_after: Duration::ZERO,
                             next: "",
                         }),
                         _ => Ok(PollerResult::Done { response }),
@@ -993,7 +1120,7 @@ mod tests {
         };
 
         let mut poller = Poller::from_callback(
-            move |_| {
+            move |_, _| {
                 let client = mock_client.clone();
                 async move {
                     let req = Request::new("https://example.com".parse().unwrap(), Method::Get);
@@ -1012,7 +1139,7 @@ mod tests {
                         match test_status.status() {
                             PollerStatus::InProgress => Ok(PollerResult::InProgress {
                                 response,
-                                retry_after: Some(Duration::ZERO),
+                                retry_after: Duration::ZERO,
                                 next: "",
                             }),
                             _ => Ok(PollerResult::Done { response }),
@@ -1083,7 +1210,7 @@ mod tests {
         };
 
         let poller = Poller::from_callback(
-            move |_| {
+            move |_, _| {
                 let client = mock_client.clone();
                 async move {
                     let req = Request::new("https://example.com".parse().unwrap(), Method::Get);
@@ -1098,7 +1225,7 @@ mod tests {
                     match test_status.status() {
                         PollerStatus::InProgress => Ok(PollerResult::InProgress {
                             response,
-                            retry_after: Some(Duration::ZERO),
+                            retry_after: Duration::ZERO,
                             next: "",
                         }),
                         PollerStatus::Succeeded => {
@@ -1185,7 +1312,7 @@ mod tests {
         };
 
         let poller = Poller::from_callback(
-            move |_| {
+            move |_, _| {
                 let client = mock_client.clone();
                 async move {
                     let req = Request::new(
@@ -1203,7 +1330,7 @@ mod tests {
                     match operation_status.status() {
                         PollerStatus::InProgress => Ok(PollerResult::InProgress {
                             response,
-                            retry_after: Some(Duration::ZERO),
+                            retry_after: Duration::ZERO,
                             next: "",
                         }),
                         PollerStatus::Succeeded => {
@@ -1307,7 +1434,7 @@ mod tests {
         };
 
         let poller = Poller::from_callback(
-            move |_| {
+            move |_, _| {
                 let client = mock_client.clone();
                 async move {
                     let req = Request::new("https://example.com".parse().unwrap(), Method::Get);
@@ -1322,7 +1449,7 @@ mod tests {
                     match no_body_status.status() {
                         PollerStatus::InProgress => Ok(PollerResult::InProgress {
                             response,
-                            retry_after: Some(Duration::ZERO),
+                            retry_after: Duration::ZERO,
                             next: "",
                         }),
                         PollerStatus::Succeeded => {
@@ -1393,7 +1520,7 @@ mod tests {
         };
 
         let mut poller = Poller::from_callback(
-            move |_| {
+            move |_, _| {
                 let client = mock_client.clone();
                 async move {
                     let req = Request::new("https://example.com".parse().unwrap(), Method::Get);
@@ -1408,7 +1535,7 @@ mod tests {
                     match test_status.status() {
                         PollerStatus::InProgress => Ok(PollerResult::InProgress {
                             response,
-                            retry_after: Some(Duration::ZERO),
+                            retry_after: Duration::ZERO,
                             next: "",
                         }),
                         _ => Ok(PollerResult::Done { response }),
@@ -1479,7 +1606,7 @@ mod tests {
         };
 
         let poller = Poller::from_callback(
-            move |_| {
+            move |_, _| {
                 let client = mock_client.clone();
                 async move {
                     let req = Request::new("https://example.com".parse().unwrap(), Method::Get);
@@ -1494,7 +1621,7 @@ mod tests {
                     match test_status.status() {
                         PollerStatus::InProgress => Ok(PollerResult::InProgress {
                             response,
-                            retry_after: Some(Duration::ZERO),
+                            retry_after: Duration::ZERO,
                             next: "",
                         }),
                         PollerStatus::Succeeded => {
@@ -1587,7 +1714,7 @@ mod tests {
         };
 
         let poller = Poller::from_callback(
-            move |_| {
+            move |_, _| {
                 let client = mock_client.clone();
                 async move {
                     let req = Request::new("https://example.com".parse().unwrap(), Method::Get);
@@ -1602,7 +1729,7 @@ mod tests {
                     match self_status.status() {
                         PollerStatus::InProgress => Ok(PollerResult::InProgress {
                             response,
-                            retry_after: Some(Duration::ZERO),
+                            retry_after: Duration::ZERO,
                             next: "",
                         }),
                         PollerStatus::Succeeded => {
@@ -1699,7 +1826,7 @@ mod tests {
         };
 
         let mut poller = Poller::from_callback(
-            move |_| {
+            move |_, _| {
                 let client = mock_client.clone();
                 async move {
                     let req = Request::new("https://example.com".parse().unwrap(), Method::Get);
@@ -1714,7 +1841,7 @@ mod tests {
                     match self_status.status() {
                         PollerStatus::InProgress => Ok(PollerResult::InProgress {
                             response,
-                            retry_after: Some(Duration::ZERO),
+                            retry_after: Duration::ZERO,
                             next: "",
                         }),
                         PollerStatus::Succeeded => {