+++ sdk/keyvault/azure_security_keyvault_certificates/CHANGELOG.md
@@ -11,6 +11,7 @@
 - `CertificateClientListCertificatePropertiesVersionsOptions::method_options` is now `PagerOptions`.
 - `CertificateClientListDeletedCertificatePropertiesOptions::method_options` is now `PagerOptions`.
 - `CertificateClientListIssuerPropertiesOptions::method_options` is now `PagerOptions`.
+- Changed `PollerOptions::frequency` from `Option<Duration>` to `Duration`.
 - Removed `CertificateClient::begin_create_certificate()`.
 - Removed `CertificateClient::resume_create_certificate()`.
 - Removed `wait()` function from `Poller<CertificateOperation>`.